The Parsoid ConfigRegistry appears to be unused
We use $wgParsoidSettings from the main registry, and never use our own
ConfigRegistry instance.  Since we eventually intend to merge the
extension code into core anyway, it doesn't really make sense to create
our own ConfigRegistry, we'll just keep using the main ConfigRegistry.

Now that we are doing this properly, with defaults in extension.json
instead of via an adhoc exception handler in the 'ParsoidSettings'
service, we can eliminate that service and simplify a bunch of code
(which, to be fair, was only recently added in
I435bdbba7d400b4226c3296cbda238dc0954edbc).

Change-Id: I0156effa3c441bdb34054832bcb8dec7a0a004f6