Move parser tests to tests/parser so they are run by default in CI
This uses the standard test runner in mediawiki-core to run any parser tests
included in tests/parser with the core/legacy parser.  This allows us to
retire the custom parsoidsvc-parsertests-docker job which manually
checked out core to do this test.  (The purpose of this test is to ensure
that we don't make local changes to Parsoid's copy of the parsertests which
would break mediawiki-core, and so avoid problems during parsertest sync.)

Change-Id: Ieac86dab494cdb90dc45daa3e02adf79637d4915