Fix integrated phan testing when optional langconv component is not present
In development langconv could be installed in Parsoid's vendor
directory but not core's.  In WMF CI, langconv is included in the
integrated vendor directory but *not* in the Parsoid standalone
modules.  So check both places.

The langconv component is optional, however; if it can't be found
either in core or in our local vendor directory, omit some classes
from phan which can't be checked properly without langconv.
(Alternatively we could provide stubs, but it's easier just to skip
the checks.)

Change-Id: Ic63a63254702e28e8128a23351a6fbe1de29f328