Move knownFailures into the Test object
This refactors the knownFailures handling to include it in
TestFileReader (so it can be reused by core) and to store the known
failures as a property of the Test object.

This does require reindexing the known failures JSON so that selser
results are stored as an alternate mode under the main test title,
instead of as a separate test title. But the result simplifies
handling: we can just pass around the Test object, instead of having
to separately pass around a knownFailures array.

In the process moved some of the statistics and known-failure-tracking
code into printResult so that we don't have to duplicate it in the XML
and non-XML reporters.  Fixed some minor bugs in the XML report output
in the process.

Change-Id: I93a4ae1b8ab46b7113f02389df7b29c98dfe0b0a