Split tools/TestUtils between new ParserTests\TestUtils and parserTests.php
Some of the code in TestUtils is specific to the command-line
parserTest runner, so it has been moved into parserTests.php.  Other
parts of the code (primarily output normalization) is needed by the
parser test runner in core, so move it to
Wikimedia\Parsoid\ParserTests\TestUtils where it can be exported.

In the process, also made the ConsoleColor class optional, since it
is in our require-dev but not in mediawiki-vendor.

Change-Id: I2e0ac8589ac99ef3a975a6f020441029eb48948f