Improve ParserTest mock of TimedMediaHandler properties
This fixes incorrect video thumb URLs in the parserTest mock: there is
an extra seek parameter, or two dashes in a row where it is omitted,
as the html/php clauses show.

Use correct mime types for audio/video (not "application/ogg"), again
to match the correct values shown in the html/php sections.

Also fix the starttime/endtime hashes, and correct the duration value
returned by the mock to match that returned by MockOggHandler in the
TimedMediaHandler extension.

Tweaked the starttime/endtime test to use offsets shorter than the
file duration, so we they are not clipped to duration and we can see
that we are actually parsing and emitting them correctly.

Bug: T258767
Change-Id: I1a116d8e9d7476e6589dee624a5c0da675d507c3