Fully implement 'page' parameter for paged media; mock it as well
We were omitting the 'page' information in the cases where width
wasn't explicitly specified in the wikitext.  Comparison to the output
of the integrated parser shows improvements to be made.

In integrated mode we weren't passing the $page parameter to
File::getWidth() and File::getHeight().  In standalone mode we need to
repeat an API call in some cases to fetch the page thumbnails after we
know the page width.

In addition to better supporting paged media, this also brings
closer the output of integrated and standalone parser test modes.

Change-Id: I11e299b987a8d8a34dcfd9a4352ab93e80a48db3