Fix O(N^2) behavior parsing nested broken templates
It turns out that the templatedepth peg parser variable was breaking
peg's parse cache, which is required to achieve good time complexity
for a PEG parser.  Because we were keeping templatedepth as an integer,
every time we resolved a broken template the new templatedepth for
all parses after that point was decremented, invalidating everything
in the parse cache.  We can't easily increment templatedepth on the
broken-template path, since we deliberately wrote those productions
so that broken templates are *not* kept on the parse stack.

In any case, we don't need to enforce templatedepth here, since we
enforce it during template expansion, and there's no longer a
time complexity reason to try to limit it during parse.  We do
predicate certain productions on whether we are 'in a template'
or not, so replace the integer templatedepth with the boolean
'intemplate' parser variable.

Bug: T259677
Change-Id: Iac96c11baeb2ed22272f29955f0a6e0ecf568657