Fix RT regression when page parameters are used with media
We were stripping the page suffix to the URL in some cases during
html2wt, but we weren't actually rigorously handling the page
parameter.  (And the way we were doing the stripping was fragile and
broke when the link title had spaces in it.)

Properly analyze the presence/absence of a `page` query suffix to the
figure link URL, and handle the cases where it is template generated
or modified by the editor.

For backwards compatibility with existing RESTBase content, don't
remove the page parameter from the wikitext when the page parameter is
missing from the href -- but we'll eventually revert this hack
once the RESTBase cache has expired (T259931).

Follow up to I11e299b987a8d8a34dcfd9a4352ab93e80a48db3.

Change-Id: I42783ba1cd0e4ebe388499d6ae7b4df5015f4ac8