Clarify uses of legacy Parser objects
Parsoid uses legacy Parser objects for two distinct uses, which are
separated and clarified by this patch:

1) A Parser object as a container for ParserOptions and the LRU cache
of fetched revisions.  This parser is privately held by PageConfig and
all revision fetches by title should go through this cache, via
Parser::fetchCurrentRevisionRecordOfTitle() (see
I1ef258252eafb63739f560039f4ead892b889222).  Two caveats:

a) When fetching a revision by ID we use a RepoStore from inside
DataAccess::fetchPageContent(); it's possible this should instead go
through a method of PageConfig for consistency or be removed
altogether (T259910).  This is actually two uses conflated together:
our initial revision fetch via the REST API is by-ID, but after that
wikitext solely uses by-title fetches to evaluate templates; etc.  The
by-title fetches can be redirected by ParserOptions in order to
implement FlaggedRevisions and other future features like coherent
time-travel and translated transclusions.

b) We'd like to eventually move this cache either into ParserOptions
or else into PageConfig (copying code), so that eventually we won't
need a Parser object to hold the ParserOptions and the cache.  As such
we are careful not to directly expose our private PageConfig::$parser
object so that no one will notice if/when it becomes unnecessary.

2) In order to invoke the legacy parser for preprocessing, PST, and
extension tag handling; and to maintain persistent state for this
purpose.  This is held by DataAccess, but the parser options for this
parser are set/reset on each invocation from the PageConfig.

We push parser object construction down the call tree by passing
ParserFactory instead of Parser, so that outside code can't see/mutate
the private Parser objects we hold for this purpose.

Bug: T236812
Change-Id: I36022385c034b0aa7e946fef3107d0cb0b093654