Make DataAccess::fetchTemplateSource() use ParserOptions::getTemplateCallback()
We were using the ParserOption revision record callback, but there's
actually a template-specific callback as well, which is what we should
have been using.  This invokes Parser::statelessFetchTemplate() which
allows extensions to hook via 'BeforeParserFetchTemplateAndtitle'
(sic) and 'ParserFetchTemplate'.  These hooks are used to customize
template resolution, for example by the Translate extension and
FlaggedRevisions.

We also move the lookup in PageConfig::getRevision() into the
PageConfigFactory so it's not repeated multiple times if the page
doesn't exist.  The implementation in PageConfigFactory now matches
the methods used by Parser::fetchCurrentRevisionRecordOfTitle(), which
is a different pathway than used for template fetch.

Change-Id: Ifb27065d3d404e45a51a15cf2bd072cc8fa1eb63