Use entity list from RemexHtml
This matches core patch Ia08bc261c3644f83109f13df04b692101b4e8ef2 and
reduces code duplication by eliminating multiple copies of the entity
list.  This also expands the set of entities accepted in wikitext to
nearly match HTML5. (HTML5 allows some entities which are not
semicolon-terminated; wiktext insists on the semicolon.)

Clean up one more edge case where we were using PHP's built-in
html_entity_decode as well, so that we have a single entity list used
everywhere. (This also fixes decode of the wikitext-specific entities
for `&rlm;`.)

Note that Parsoid had generally already accepted the full HTML5 entity
list, as a result of using html_entity_decode() instead of
Sanitizer::decodeEntity().  This was a difference between the legacy
parser and Parsoid which has been addressed by the core patch.  A
parser test illustrated this difference; it used &vert; which was one
of the entities accepted by Parsoid but not by core.  A variant was
added to the parser test using &#x7C; instead, which was recognized by
both parsers. (Note that the legacy parser still has a bug in parsing
alt text in <gallery>; that's T49646.)

Bug: T94603
Depends-On: Ia08bc261c3644f83109f13df04b692101b4e8ef2
Change-Id: I873a6120e4bd1c69fee9da76d266e24e97a22add