Convert error Response returns to throws of HttpException
This makes it more obvious that request processing is intended to
immediately halt when these conditions are encountered, and avoids the
easy mistake of forgetting to check for an error condition.

In the process also restructured respondToMissingRevisionContent() to
tryToCreatePageConfig().  The new form is more intuitive since it is
clear we are trying to create PageConfig and might encounter some
exceptional conditions (which throw, so we don't need to explicitly
check for them).  It also eliminates the create-by-reference of
$pageConfig which is pretty confusing wrt reading code.

Throwing to handle redirects is T260959 upstream; we've hacked around
it temporarily in our code base by creating our own ResponseException
wrapper object.

Change-Id: I800a3fe5160a9d7fc3fddbb445ec61cc5390b14f