PageConfigFactory: don't use latest revision if getRevisionById() fails
In situations with replica lag, getRevisionById() can fail even if the
revision is present in the master DB.  Previously Parsoid would
transparently fetch the 'latest' revision instead, which could lead to
corruption during selser.

Fix this problem with belt and suspenders: first, explicitly try the
master DB if getRevisionById() fails. This should be a very rare
occurrence: when Parsoid is supplied with a revision ID it is *always*
assumed to exist.

Second, tweak the code paths to ensure that even if getRevisionById()
fails we never silently fetch the latest revision instead.

We also restructured the $wikitextOverride and access permission
checks to ensure they apply uniformly no matter how we got our
$revisionRecord.

Bug: T259855
Change-Id: I7fa7fb05636180cd3cf2dcd25c072363aeaa4563