Sometimes oldid is '' instead of null
It turns out that to work around T232556, we added a trailing slash to
urls in roundtrip-test.js in 3d193c9b. So, as a result, it appears that
the REST code treats the oldid component as ''.

During porting, by casting the oldid value to bool, we effectively worked
around this issue without knowing it and tripped over it when we removed
that ugly casting in 7fe64ea5.

T232556 is a result of IE6/IE7 support which was dropped as part of
T232563. So, we should be able to remove our T232556 workarounds in the
codebase. But for now, let us unbreak rt-testing by handling the ''.

Change-Id: I4b2d5c68da7524a9d112c47734b03407ed24084c