Remove special case for the html extension when unpacking
The special case for the html extension (d40afe4 / T185266) is ripped
out since it was unnecessary after the port.  The validation issue when
cloning nodes was specific to domino and no longer applies,
https://github.com/fgnass/domino/blob/12a5f67136a0ac10e3fa1649b8787ba3b309e9a7/lib/Element.js#L315-L327

However, in 4227d70 / T227209, we added a client error for when we're
given HTML to serialize that has names that are invalid in XML.  That
validation was broadly being applied to all HTML we parse, including on
the content of html extension tags in the parse direction.  Here we
restrict that to only validating HTML at the entrypoint.  In the parsing
direction, names are sanitized or coerced where appropriate.

Bug: T221790
Bug: T179082
Bug: T217705
Bug: T274155
Change-Id: Ia229f6ea254f1764453d9489924907831d145afe
(cherry picked from commit 19a9c3ce9e037dfb733a88d25079d64753897cc9)