Avoid array_key_first() for compatibility with PHP 7.2
It's new in PHP 7.3:
https://www.php.net/manual/en/migration73.new-functions.php

Current MediaWiki is still compatible with PHP 7.2.22, and we
use PHP 7.2.31 in Wikimedia producion. The only reason it's not
failing there is because apparently some extension loads the
Composer module 'symfony/polyfill-php73'.

In this particular situation, the array in question always has
exactly 1 element, so array_key_first() has minimal performance
advantage over array_keys()[0].

Bug: T261044
Change-Id: Ide2beae3c26706ed2efc6ff2dd7ccbfece5bfd86