Handle media-in-link scenario
Follow up to 85be08e

This was hidden previously by the check for already processed media and
exposed when that was removed.

WikiLinkHandler::addLinkAttributesAndGetContent() drops all the anchor
tokens it sees, which messes with the expected structure in the
AddMediaInfo processor.  Instead, we bail early on wikilink with media
in its contents, to match core output.

Fixes `echo "[[Test|[[File:Test.png]]]]" | php bin/parse.php`

Change-Id: Iedf9141f1510043d9cbc633d0926a227c5ac8c5a