Preserve typeof on extension first encap wrapper nodes
The imagemap extension surfaced this since it's desirable to have media
typeof on the figure, which is the only node returned.

However, there may be negative consequences to doing this if the first
node returned is somehow considered template generated, since templates
have a higher precedence, the about extents may not line up.

Also, similar to template wrapping, the data attribs are clobbered but
that isn't resolved here.

Bug: T214241
Change-Id: I95767e466803f0744b6626204a0a3a1514fff174