TableFixups: Simplify combinable cell checks
All we need to check is that the cell is a <td> and that
a previous node exists and that it is an element.

If the previous node is a comment / whitespace, it would be
the first node of the current source line and there is no
opportunities to combine cells in that case.

This wikitext below

  {|
  |-
  <!--c1-->|foo1
  <!--c2-->|foo2
  |-
  <!--c3-->|foo3
  |}

parses as:

  <table>
  <tbody><tr>
  <!--c1--><td>foo1</td>
  <!--c2--><td>foo2</td></tr>
  <tr>
  <!--c3--><td>foo3</td></tr>
  </tbody></table>

Change-Id: I8ccf196c25e956e01c3cf4f42d2ef23d5774b7f8