Handle mw:DisplaySpace when serializing mw:Nowiki content
Follow up to 275cb60 that exposed it but this was a problem since
9fd3efe landed.

Also, adjust normalizing mw:DisplaySpace on the JS side to match PHP
after d88895d, since it's still used in rt.

Test is added which fails to roundtrip on master.

The fate of whether mw:DisplaySpace should be applied inside of
mw:Nowiki is still to be decided in T255007.

Bug: T255007
Change-Id: Iab9c4255c511f4fb3aa59f62011faad5d3501e04