build: update mediawiki-phan-config to 0.10.3 and fix TMPDIR use
Without php-ast, it causes T251346:

    Error: Undefined constant 'ast\AST_LIST'

That surfaced on CI since Quibble 0.0.45 erroneously run composer test.
Something that should be addressed in a future release of Quibble.

Quibble is also setting TMPDIR now (apprently it didn't use to?) and
is setting it to `/tmp` (no trailing slash) which is breaking our
toolcheck script.  Although that's a regression from earlier quibble
releases, that's argubly our bug: I shifted the slash in our script so
that it works whether or not your TMPDIR ends in a slash.

Bug: T263500
Change-Id: I3196aff663848ce2dc7c2e5498832d4a23d4633f