Skip tests/phpunit when running 'extension unit tests' from core
The tests in tests/phpunit are the tests for the standalone Parsoid
library.

If we have any integrated-mode unit tests to run as part of the
Parsoid extension we'll put them in extension/tests/phpunit.

Note that tests/parser/*.txt *are* supposed to be run in both
extension mode and standalone; this is the way we ensure that
we don't add `html/php` sections to parser tests which don't
pass when run in the legacy parser.

Change-Id: I3b91cb6576e6d00836f88b987a262f7faa3a4c15
(cherry picked from commit c79a13bf2a30b27198774903add3cd1c27b76446)