Skip tests/phpunit when running 'extension unit tests' from core
The tests in tests/phpunit are the tests for the standalone Parsoid
library.

If we have any integrated-mode unit tests to run as part of the
Parsoid extension we'll put them in extension/tests/phpunit.

We also need to add 'extension:TimedMediaHandler' to the '!! hooks'
section of tests/parser/timedMediaHandlerParserTests.txt to prevent
it from being run in CI when core doesn't have TimedMediaHandler
loaded.

Note that tests/parser/*.txt *are* supposed to be run in both
extension mode and standalone; this is the way we ensure that
we don't add `html/php` sections to parser tests which don't
pass when run in the legacy parser.

Depends-On: I2d3f7e1313b4456733f820e6d8c504bb8d7427a7
Change-Id: I3b91cb6576e6d00836f88b987a262f7faa3a4c15