Convert offsets in pb2pb after data attribs are loaded
In WikitextContentModelHandler and DOMPostProcessor, similar calls to
convert offsets are made while data attribs are loaded.  This change is
made in order to consistently call the method when the dom is in the
same state and avoid any potential bugs if data attribs are used.

This is similar to the changes made for redlinks (3888354) and langconv
(1641a20).

Change-Id: I4a88186af65137aea38ba46419f6d55182f59c25