Move whitespace heuristics for trimmable whitespace nodes to buildSep
T157418 made whitespace trimming mandatory for list items, table cells,
and headings. This then necessitated heuristics in the HTML -> WT
direction to insert any necessary trimmed whitespace to reduce dirty
diff incidents.

As observed in T263502, it is better for these heuristics to be
combined with the separator heuristics in Separators::buildSep.

This match effects that change and also updates the heuristics and
cleans up that code as well.

1. Since we expanded DSR information with two more elements to capture
   length of trimmed leading and trailing whitespace, we can leverage
   that information where available before using any heuristics.

2. Cleaned up the conditions for where the heuristics apply. We don't
   need to check for whether the node is new or not.

   For leading whitespace, all we care about is checking if its content
   starts with whitespace or not (we don't need to apply heuristics
   if content starts with whitespace).

   The checks for trailing whitespace are equivalent.

However, since we moved these heuristics into buildSep, the heuristics
will not kick in unless separators are emitted on both opening and closing
tags of nodes. So far, we didn't invoke buildSep (via emitSepForNode)
for <td>, <th>, <li> since those nodes don't have closing tags in wikitext.
This patch fixes that.

After a bunch of experimentation, decided to recover these trailing
separators only on unmodified nodes.

Test Result Changes
-------------------
* Most selser test result changes look like improvements.
* However, there are some cases where on mass subtree modifications,
  the list-handler is recovering trimmed whitespace from one of those
  deleted list items. It is questionable as to whether this is desirable.

  Ex: Test "Nested lists 11 (T262409: trailing whitespace & selser)"
  with selser changetree code of "[[[4]]]"

  Instead of adding more complexity to this patch, we'll let the
  followup patch take care of it since it seems to be fixing up a number
  of these tests.

Bug: T263502
Change-Id: I95b841370efbd37c4acc6137216bafd83912e612