Protect templated tokens from running through the ae multiple times
Follow up to 28a7481

In 28c5449, it was established that tokens passing through the attribute
expander haven't been expanded to dom yet so it is safe to unpack any
dom fragments encountered when converting tokens to string.

However, one of the jobs of the attribute expander is to expand values
to dom and hence after passing through the attribute expander the above
is no longer true.

When template expansion happens, text is processed to the end of
stage-2, which is post-attribute expansion.  However, those tokens are
then reinserted in the stream at the beginning of stage-2 where
template expansion occurs and they go through stage-2 a second time.
That's mostly harmless because in the first go around token types change
(ex. wikilink to anchor) so the stage appears idempotent.  But attribute
expansion didn't have that sort of protection.

Change-Id: I6280a8a2a6fe40ad2104d737d0cd9171497b7aa5