Add a script to kick off rt testing
We run this script locally. It updates code on scandium and
testreduce1001 and kicks off rt testing.

The code update on testreduce1001 is optional and only matters
where rt-testing related code is updated. But, it is simpler to
just update it every single time.

Since the contents of the update_parsoid.sh script has been
inlined here, we can get rid of it from the puppet repo after
this is merged and tested.

Bug: T257906
Change-Id: I45f6ee6ba60d364b8c7a6a7df43826384c38248d