Bump wikipeg to a fork of 2.0.4
And regenerate the grammar.

This adds back the changes that were reverted in 5eabdc5.

Since 870be90c, we've been generating the grammar with this new
unreleased version.

Unfortunately, there was a mistake in the commit that made the bump /
tag, in that, it never incremented the package.json so we'll need to
move to a 2.0.5

Change-Id: I9bc76c36254bb895db9094001b9105768ea99c02