Templated template targets are already expanded to dom
At the top of onTemplate, if we detect that the template target is
templated, we send the token through the attribute expander, which
expands attributes to dom.

So, when template expansion is aborted, convertAttribsToString puts
tokens back in the stream that can potentially already have been
unpacked.

Rather than trying to get fancy about this scenario, just use the
template source.

A regression test is added that throws on master.

Bug: T266171
Change-Id: I6ceff9a8562919ff91065b30c0bbca03cba824cc