Fix processing embedded keys in processHiddenHTMLInDataAttributes
It was only looking at values, not keys, which could also have html
components.  The result was that sealed fragments were leaking in the
embedded content and throwing when they were encountered on the way back
in (html2wt) when trying to shift the dsr.

Note that, while this fixes the leaking for future renders, content in
our storage will still throw when being asked to serialize.

This implementation matches ContentUtils::shiftDSR, which probably means
we should have some consolidated method for doing this traversal.

Bug: T266171
Change-Id: Iec972caf15628412d41b6f7954111b3b51990b43