phan config: no more unported files; use file list instead of regex
Remove the no-longer-needed code to exclude unported files from phan.

Use the wfCollectPhpFiles function to avoid the need for
exclude_file_regexp, which is brittle: if mediawiki-phan-config were
to set exclude_file_regexp there's no good way to "append" to it;
currently we'd just overwrite the upstream setting.  Instead just
collect all the files from the directories we want to exclude and
use exclude_file_list.

Change-Id: Idf3ecfb5c7f4b171d1f82d3f70fd22ac0dafe97a