Split up ContentUtils::ppToDOM to avoid document creation
This adds ::createAndLoadDocument and ::createAndLoadDocumentFragment
and removes ::ppToDOM

Throughout parsing and serializing, we shouldn't ever need to create
documents and will instead turn to creating and loading fragments.

The other method is added to create documents with data attributes loaded
for all the remaining places that were still using ppToDOM to create a
document, which are all during testing.

Testing documents don't necessarily need to be retained with the live
docs on the environment since they can remain in a variable throughout
the test and the bag won't be collected.

Bug: T265061
Change-Id: I74ac9e5f9cfcd0c7d3dec33bf067d605b51d95e4