Shore up serializing spans for inline media
This makes the checks for inline media throughout a little more generic.

The shoring up is mainly because the check for missing media (rendered
as a redlink) was hardcoded to start with ^figure.  This shouldn't
really be a blocker to editing clients starting to send us inline media
in spans, because presumably newly created ones won't be for missing
media.

There's a test to show we still accept spans for inline media,
"Serialize simple image with span wrapper"

Bug: T266143
Bug: T251641
Change-Id: Id9b680ef0daf5dad4c41b0ecf392d12feed6f4fe