Selser: Fix incorrect recovery of trimmed whitespace
Should not attempt recovery for intermediate span wrappers.

Added a regression phpunit test (since it is harder to add a
parser test because of inability to target text nodes via
CSS selectors).

It looked like a fix of T266908 would have fixed this as well
but it turns out it caused failure of an existing parser test.
Removed the FIXME as invalid and documented the reasons if
someone else is templated to do the same in the future.

Added phpdocs on a couple of associated functions for improving
code readability.

Bug: T268737
Change-Id: I04a637a734058cde1a43e04865a2572fac2fba26