ParserTestReader: Support a generalized file-level options clause
This generalizes the existing !!hooks and !!functionhooks mechanisms
and allows a top-level !!options clause at the top of the file to
specify file-level options.  This will allow opt in/opt out of various
test modes as well as specifying dependencies on specific extensions,
etc.

In the near term this will be used to opt in to running extension
parser tests using Parsoid in integrated mode.

Pull the !!version clause into this new !!options mechanism as well,
so we're dealing with all file-level options consistently.

End the !!options section at the top with !!end (rather than
introducing a new !!endoptions flag) and make this consistent by
allowing !!end as an alternative to !!endhooks, !!endfunctionhooks,
and !!endarticle.

In summary, this patch will allow the parser test files to start like:

!! options
version=2
parsoid-compatible
requirements=references,{"type":"hook","name":"ref"},extension:Gallery
!! end

Bug: T254181
Change-Id: I636bd1f2c8aee327acbbd1636e2ac76355f1d80e