Tokenizer: Don't eat leading spaces from template values
Most template and parser function invocations strip both keys and values
of template arguments, but that's not an inherent limitation -- some
parser functions may wish to (eventually) opt out of that behavior,
so they don't have to play <nowiki> games to preserve meaningful
whitespace (T204307).  But even before that point, stripping whitespace
in the tokenizer leaves us in a pickle when the template name turns
out to be invalid and we need to reserialize the keys and values.

Change-Id: I6430f8edc9748e5a4058e2ad6da64bf88d990166