Delint two issues flagged by phan when run in PHP 8.0
The null-coalescing after json_encode is unnecessary because the
argument is declared as type `array` in the method signature.  An
array will always json_encode correctly, and if the parameter is not
an array it will throw an exception when the method is called.

Similarly, the result of base_convert() should always be a string,
and if it were not, the callback is declared to return a string so
it will be cast to string anyway.

Bug: T269719
Change-Id: I7905864b1eaef15243641c7e9dd66693969ba0ad
(cherry picked from commit 2e2b073f8937ad6f047e3ea3d027077595e3926a)