cced5578 followup: Handle newlines in wikilinks for selser as well
cced5578 only handled this for when the wikilink was serialized
by the regular serializer, but forgot to handle the scenario when
the original source for the wikilink went through selser and got
modified to suppress newlines to avoid breakage of surrounding
context (lists, etc.).

Added new test that specs this behavior.

Bug: T265094
Change-Id: Id476a9f63aaff26ad33fc58418fd62c055016429