Modify UTF-8 regex to use builtin PCRE validation
The existing regex was causing PCRE to fail for very long articles due
to hitting pcre.recursion_limit. PCRE has a built-in UTF-8 validator
exposed via the 'u' modifier, see
https://www.php.net/manual/en/reference.pcre.pattern.modifiers.php for
more information.

The 'u' modifier causes the match to fail if the subject is invalid
UTF-8 regardless of what the regex pattern actually is. So, with an
empty regex pattern we will always succeed on valid UTF-8 strings and
always fail on invalid UTF-8 strings. Furthermore, as this is part of
PCRE, it doesn't require backtracking or recursion in order to validate
the string.

Bug: T272232
Change-Id: I4a4060570472f8223eecf3e7d6d0976fbcd0ccca