Don't apply border class to thumbs
Note that roundtripping this option depends on the class being added
(serialization is taken from the class) so "|border" will get dropped
from thumbs and we'll need to rely on selser not to dirty diff.

Alternatively, we can use the bogus attribute mechanism to preserve it
where it isn't applicable.

This matches the legacy parser because "thumbborder" doesn't get added
in makeThumbLink2, meaning that it doesn't apply to framed/thumb media.

No changes to parserTests because only the applicable cases are tested.

Change-Id: I54e451e2bf30a1ebab4c0611e2f4dcf191f0c43f