Handle wikilinks misnested in media links
In the legacy parser, nesting (wikilink-in-wikilink) is only matched
when the outer link has an NS_FILE prefix.

In Parsoid, this is enforced in the WikiLinkHandler by bailing on the
wikilink when nesting is detected in its content, see a2de7d2.

MediaLinks had their own adhoc content handling, instead of using
addLinkAttributesAndGetContent, so didn't have this detection.

A test is added for the bug and another to show that tokens with
multiple maybeContent were dropping some of them on the floor.

Bug: T278074
Change-Id: I58eab4554db3d890e641418b92ce30c591996eba