Don't p-wrap <aside> tags in extension HTML
Our PortableInfobox extension uses the HTML5 <aside> tag in its generated HTML.
Parsoid attempts to p-wrap this tag, which results in superfluous empty
paragraph wrappers in the wt2html output, causing dirty diffs when an article
containing such an extension tag is edited in the VisualEditor.

As a fix, ensure Parsoid doesn't try to p-wrap <aside> tags, similar to how it
treats <figure>. Like <figure> , this patch allows <aside> to exist in
extension or core-generated HTML, but will sanitize it away if the tag pops up
in raw wikitext.

Bug: T278565
Change-Id: I3e57f55ac69d2c1ee8a1d41c21b692e56fc7e628