Handle optional spaces after table_attributes for table_row_tag
Without this, full_table_in_link_caption fails so the table in the
example doesn't parse.

The changes to table_row_tag are to match table_start_tag but maybe
there's a more general refactor here since we seem to have to be careful
about spaces trailing table_attributes in all the places it's used.

Bug: T279682
Change-Id: Iba9ac1908c858741f1a93430a901a14d422ddfbd