Account for nested transclusions in table fixups
Follow up to e8bcdd1.  The transclusions array can now contain nodes
that are not direct descendants of the td.

This is further fallout from the algorithm in e8708ad not accounting for
the contents of dom elements found in attribute positions.  A FIXME is
added to highlight further edge cases that can be contrived.

A test is added that throws as reported in the task.

Bug: T280672
Change-Id: Ia534b81a3a2625acc050e1921ec5be7e6b797e71