Pass media structure to figure handler
Add a MediaStructure class and pass instances of it to
LinkHandlerUtils::figureHandler instead of trying to parse all that info
at the top of it.  This puts the burden on the callers who have better
insight anyways.

These changes lead to a little bit less compatibility in serializing
arbitrary html, as opposed to Parsoid flavour html, so some decisions
need to be made about what we want to support.  For example:

The blacklist change raises the question of how hard we should try to
find the media element within the wrapper.  Previously we used a query
selector but now we're just looking at the first non-seperator child.

Change-Id: I6055f04d1a1e00101cd96cb97a2325a182588ba7