Enforce wikitext limits like in the legacy parser
MediaWiki enforces $wgMaxArticleSize when editing an article and
so there doesn't seem to be a need to check it at parse time.  We keep
the size limit at the entrypoint because Parsoid also receives arbitrary
wikitext posted to it, not just wikitext coming from articles in the db.

The legacy parser does two checks for maximum inclusion size though
(which is also set to $wgMaxArticleSize).  It does a fast check to see
if we're already at the max when the parse begins, in which case
template expansion is skipped altogether.  Otherwise, it keeps track of
the expansion size of templates and makes sure the accumulation is less
than the max inclusion size.  If it's reached, template expansion halts.

Note that the legacy parser keeps a separate count for template
argument expansion, which isn't implemented here.

Limits are adjusted to match $wgMaxArticleSize and the comparison
switches from mb_strlen to strlen to match core.

Bug: T280381
Bug: T211946
Bug: T239841
Change-Id: I4d7e3816a3c37383130420f86424e38bfac1d15f