Replace freenode references with libera references
References to freenode are not yet modified in baseconfig: these files
are generated by sync-baseconfig.js, which is currently broken; this
will be addressed in a separate change. We are keeping a reference to
freenode in siteinfo.reqdata since this comes from an existing API call.

siteinfo.reqdata was updated with this patch with a fresher version
taking into account among others the addition of libera.chat.
The corresponding tests in SiteConfigTest (affecting other elements of
this file, such as the NoCollaborationHubTOC magic word) were adjusted
accordingly.

Bug: T283273
Change-Id: I3aa3a0cca39935532dda185163da9da466618837